requirements.txt
app==0.0.1
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.45.0
django-appconf==1.0.4
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-model-utils==4.1.1
django-mptt==0.13.2
django-notifications-hq==1.6.0
django-password-reset==2.0
django-redis==4.12.1
django-simpleui==2021.8.17
django-taggit==1.5.1
Flask==2.0.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonfield==3.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
oauthlib==3.1.1
pilkit==2.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2021.1
pywin32==300
pyWinhook==1.6.2
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
